backend/src/requirements.txt
chromadb
cryptography
ollama
pillow
django
django-ninja
django-ninja-extra
psycopg2-binary
django-ninja-jwt
django-cors-headers
requests
gunicorn
whitenoise
celery
redis
django-redis
django-celery-beat
pydantic[email]
dj-database-url
pywebpush
